Remove detail/invoke.hpp
The internal file 'invoke.hpp' has been removed, and implementation
moved to the public 'invoke.hpp'
